Add `bentoml containerize` command (#847)
* add basic draft containerize

* more wip stuff

* linting + formatting

* refactored to use CLIException

* fix bad api catch and better docs

* fix bad push catching and add colour

* linting

* fixed bad layer size calculation

* better unpack

* address comments

* address naming

* update containerize docs

* wip migration

* fixed bad imports

* refactor bento_service tag handling behaviour

* fix bad patch

* update makefile and properly fix patch

* add _echo_docker_api_result tests

* fixed bad comparison in test

* add basic e2e and test_make_bento_name_docker_compatible

* remove e2e for another pr
