+++ tests/cli/test_cli.py
@@ -3,10 +3,16 @@ import tempfile
 
 import pytest
 import mock
+from humanfriendly import format_size
 from click.testing import CliRunner
 import psutil  # noqa # pylint: disable=unused-import
 
-from bentoml.cli import create_bento_service_cli
+from bentoml.cli.bento_service import (
+    create_bento_service_cli,
+    make_bento_name_docker_compatible,
+)
+from bentoml.cli.utils import _echo_docker_api_result
+from bentoml.exceptions import BentoMLException
 
 
 def generate_test_input_file():
@@ -19,6 +25,34 @@ def generate_test_input_file():
     with open(file_path, "w") as f:
         f.write('[{"col1": 1}, {"col1": 2}]')
     return file_path
+
+
+def assert_equal_lists(res, expected):
+    assert len(expected) == len(res)
+    assert all([a == b for a, b in zip(expected, res)])
+
+
+def test_make_bento_name_docker_compatible():
+    names = [
+        ("ALLCAPS", "allcaps"),
+        ("...as.df...", "as.df"),
+        ("_asdf_asdf", "asdf_asdf"),
+        ("1234-asdf--", "1234-asdf"),
+    ]
+
+    for name, expected in names:
+        assert make_bento_name_docker_compatible(name, "") == (expected, "")
+
+    tags = [
+        ("....asdf.", "asdf."),
+        ("A" * 128, "A" * 128),
+        ("A" * 129, "A" * 128),
+        ("-asdf-", "asdf-"),
+        (".-asdf", "asdf"),
+    ]
+
+    for tag, expected in tags:
+        assert make_bento_name_docker_compatible("", tag) == ("", expected)
 
 
 def test_run_command_with_input_file(bento_bundle_path):
@@ -58,6 +92,96 @@ def test_run_command_with_input_file(bento_bundle_path):
 
     assert result.exit_code == 0
     assert result.output.strip() == '3'
+
+
+def test_echo_docker_api_result_build():
+    build_stream = [
+        {'stream': 'Step 1/2 : FROM bentoml/model-server:0.8.1'},
+        {'stream': '\n'},
+        {'stream': ' ---> f034fa23264c\n'},
+        {'stream': 'Step 2/2 : COPY . /bento'},
+        {'stream': '\n'},
+        {'stream': ' ---> Using cache\n'},
+        {'aux': {'ID': 'sha256:garbagehash'}},
+        {'stream': 'Successfully built 9d9918e008dd\n'},
+        {'stream': 'Successfully tagged personal/test:latest\n'},
+    ]
+
+    expected = [
+        "Step 1/2 : FROM bentoml/model-server:0.8.1",
+        " ---> f034fa23264c",
+        "Step 2/2 : COPY . /bento",
+        " ---> Using cache",
+        "Successfully built 9d9918e008dd",
+        "Successfully tagged personal/test:latest",
+    ]
+
+    res = [line for line in _echo_docker_api_result(build_stream)]
+    assert_equal_lists(res, expected)
+
+
+def test_echo_docker_api_result_push_no_access():
+    push_stream = [
+        {'status': 'The push refers to repository [docker.io/library/asdf]'},
+        {'status': 'Preparing', 'progressDetail': {}, 'id': '2e280b8a5f3e'},
+        {'status': 'Preparing', 'progressDetail': {}, 'id': 'd0b7e1b96cc1'},
+        {'status': 'Preparing', 'progressDetail': {}, 'id': 'fcd8d39597dd'},
+        {'status': 'Waiting', 'progressDetail': {}, 'id': '2e280b8a5f3e'},
+        {'status': 'Waiting', 'progressDetail': {}, 'id': 'd0b7e1b96cc1'},
+        {'status': 'Waiting', 'progressDetail': {}, 'id': 'fcd8d39597dd'},
+        {
+            'errorDetail': {
+                'message': 'denied: requested access to the resource is denied'
+            },
+            'error': 'denied: requested access to the resource is denied',
+        },
+    ]
+
+    with pytest.raises(BentoMLException) as e:
+        _ = [line for line in _echo_docker_api_result(push_stream)]
+    assert str(e.value) == 'denied: requested access to the resource is denied'
+
+
+def test_echo_docker_api_result_push():
+    push_stream = [
+        {'status': 'The push refers to repository [docker.io/personal/test]'},
+        {'status': 'Preparing', 'progressDetail': {}, 'id': '2e280b8a5f3e'},
+        {'status': 'Preparing', 'progressDetail': {}, 'id': '03613b6b1004'},
+        {'status': 'Waiting', 'progressDetail': {}, 'id': 'cb2960c4c4d1'},
+        {
+            'status': 'Pushing',
+            'progressDetail': {'current': 5632, 'total': 532223},
+            'progress': '[=>    ]',
+            'id': '03613b6b1004',
+        },
+        {
+            'status': 'Pushing',
+            'progressDetail': {'current': 512, 'total': 699},
+            'progress': '[=====> ]',
+            'id': '2e280b8a5f3e',
+        },
+        {
+            'status': 'Pushing',
+            'progressDetail': {'current': 128512, 'total': 532223},
+            'progress': '[==>   ]',
+            'id': '03613b6b1004',
+        },
+        {'status': 'Pushed', 'progressDetail': {}, 'id': '2e280b8a5f3e'},
+        {'status': 'latest: digest: sha256:garbagehash size: 2214'},
+        {
+            'progressDetail': {},
+            'aux': {'Tag': 'latest', 'Digest': 'sha256:garbagehash', 'Size': 2214},
+        },
+    ]
+
+    expected = [
+        f"Pushed {format_size(5632)} / {format_size(532223)}",
+        f"Pushed {format_size(5632 + 512)} / {format_size(532223 + 699)}",
+        f"Pushed {format_size(128512 + 512)} / {format_size(532223 + 699)}",
+    ]
+
+    res = [line for line in _echo_docker_api_result(push_stream)]
+    assert_equal_lists(res, expected)
 
 
 @pytest.mark.skipif('not psutil.POSIX')
